Added main & exception macros. Build and runs but panics after init. Is this bcause it doesnt fall into idle?
