+++ .vscode/launch.json
@@ -22,5 +22,26 @@
                 "step"
             ]
         },
+        {
+            "type": "gdb",
+            "request": "attach",
+            "name": "WatchDebug",
+            // "gdbpath" : "/home/mabez/programs/gcc-arm-none-eabi-5_4-2016q3/bin/arm-none-eabi-gdb",
+            "gdbpath": "/home/mabez/programs/gcc-arm-none-eabi-7-2017-q4-major/bin/arm-none-eabi-gdb",
+            "debugger_args": [
+                "-nx" // dont use the .gdbinit file
+            ],
+            "executable": "./target/thumbv7em-none-eabi/debug/mabez_watch",
+            "remote": true,
+            "target": ":3333",
+            "cwd": "${workspaceRoot}",
+            "autorun": [
+                "set print asm-demangle on",
+                "monitor arm semihosting enable",
+                "load",
+                "break main",
+                "step"
+            ]
+        },
     ]
 }
